Set up stdout & stderr cache files for r_bg process

Set up stdout & stderr cache files for r_bg process

logfile_names(path)

Arguments

path
Path to local repository

Value

Vector of two strings holding respective local paths to stdout and stderr files for r_bg process controlling the main pkgcheck function when executed in background mode.

Note

These files are needed for the callr r_bg process which controls the main pkgcheck. The stdout and stderr pipes from the process are stored in the cache directory so they can be inspected via their own distinct endpoint calls.

See also